Move churn attack test out of routing
